// projects/qa/src/app/models/Ticket.ts
export class Ticket {
  id: number;
  responsible: string;
  urgency: string;
  productName: string;
  project_name: string;
  shortDescription: string;
  longDescription: string;
  testId: number;
  testStepId: number;
  productId: number;
  userStoryId: number;
  targetSprintId: number;
  originSprintId: number;
}